Move OutboundCall to a separate file
